1) Refactored singular routedns client and models files into unique files per routedns feature to improve readability and maintainability 2) Added comments to most struct parameters 3) Updated return value for failed Group creation ID from -1 to nil
